+++ src/main/java/com/nortvis/demo/configuration/security/JwtRequestFilter.java
@@ -1,5 +1,6 @@
-package com.nortvis.demo.configuration;
+package com.nortvis.demo.configuration.security;
 
+import com.nortvis.demo.configuration.util.JwtUtil;
 import com.nortvis.demo.service.impl.UserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
