+++ src/main/java/com/nortvis/demo/feignclient/ImgurService.java
@@ -1,0 +1,4 @@
+package com.nortvis.demo.feignclient;
+
+public interface ImgurService {
+}
